Remove Rule92NotSystemUserHelper
