+++ src/main/controllers/helpers/StoredApplicationConfirmationHelpers.ts
@@ -1,0 +1,9 @@
+import { GenericTseApplicationTypeItem } from '../../definitions/complexTypes/genericTseApplicationTypeItem';
+import { Applicant } from '../../definitions/constants';
+
+export const getLatestApplication = async (
+  items: GenericTseApplicationTypeItem[]
+): Promise<GenericTseApplicationTypeItem> => {
+  const filteredItem = items?.filter(it => it.value.applicant === Applicant.CLAIMANT);
+  return filteredItem[filteredItem.length - 1];
+};
